+++ NEWS.md
@@ -10,6 +10,7 @@
     * `trakt.people.summary`
     * `trakt.people.movies`
     * `trakt.people.shows`
+* Unify output of `trakt.show.people` and `trakt.movie.people`
 
 ### v0.10.3